% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compensation_error.r
\name{get.compensation.error}
\alias{get.compensation.error}
\title{Get Compensation Error}
\usage{
get.compensation.error(
  expr.data.unco,
  expr.data.comp,
  marker.spillover.unco,
  scale.untransformed,
  plot.figure,
  figure.label,
  flow.gate,
  flow.control,
  asp
)
}
\arguments{
\item{expr.data.unco}{Matrix of uncompensated expression data.}

\item{expr.data.comp}{Matrix of compensated expression data.}

\item{marker.spillover.unco}{List of matrices describing uncompensated spillover.}

\item{scale.untransformed}{Logical value indicating whether to use untransformed scale.}

\item{plot.figure}{Logical value indicating whether to plot figures.}

\item{figure.label}{Character string representing the label for the figure file.}

\item{flow.gate}{List containing flow gate parameters.}

\item{flow.control}{List containing flow cytometry control parameters.}

\item{asp}{List containing aesthetic parameters for the plot.}
}
\value{
A list of matrices describing compensation error, including intercepts,
coefficients, slopes, and skewness.
}
\description{
This function returns a list of matrices describing compensation
error, including intercepts, coefficients, slopes, and skewness.
}
