% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_log_together.R
\name{plot.density.log.together}
\alias{plot.density.log.together}
\title{Plot Density of Log Values Together}
\usage{
\method{plot}{density.log.together}(x.table, x.label, plot.file.path, asp)
}
\arguments{
\item{x.table}{List of matrices containing the data to be plotted.}

\item{x.label}{Character string representing the label for the x-axis.}

\item{plot.file.path}{Character string representing the file path to save the plot.}

\item{asp}{List containing aesthetic parameters for the plot.}
}
\value{
A ggplot object representing the density plot of log values.
}
\description{
This function plots the density of log values for initial and
final steps of autospill, including calculations with positive and
negative populations.
}
