% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gate.r
\name{plot.gate}
\alias{plot.gate}
\title{Plot Gate}
\usage{
\method{plot}{gate}(
  gate.stage,
  samp,
  gate.data,
  gate.marker,
  gate.bound,
  gate.region,
  gate.population,
  flow.control,
  asp
)
}
\arguments{
\item{gate.stage}{Integer indicating the stage of gating.}

\item{samp}{Character string representing the sample identifier.}

\item{gate.data}{Data frame containing the gating data.}

\item{gate.marker}{Vector of markers used for gating.}

\item{gate.bound}{Data frame containing the boundary data for gating.}

\item{gate.region}{Data frame containing the region data for gating.}

\item{gate.population}{Data frame containing the population data for gating.}

\item{flow.control}{List containing flow cytometry control parameters.}

\item{asp}{List containing aesthetic parameters for the plot.}
}
\value{
Save a PNG file.
}
\description{
This function generates and saves a plot representing the gate
for flow cytometry data.
}
