% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.r
\name{plot.matrix}
\alias{plot.matrix}
\title{Plot Matrix Coefficients}
\usage{
\method{plot}{matrix}(the.matrix, by.rows, figure.dir, figure.file.label, flow.control, asp)
}
\arguments{
\item{the.matrix}{Matrix containing the coefficients to be plotted.}

\item{by.rows}{Logical value indicating whether to plot by rows.}

\item{figure.dir}{Character string representing the directory to save the plot.}

\item{figure.file.label}{Character string representing the label for the figure file.}

\item{flow.control}{List containing flow cytometry control parameters.}

\item{asp}{List containing aesthetic parameters for the plot.}
}
\value{
A ggplot object representing the matrix plot.
}
\description{
This function plots matrix coefficients by rows or columns.
}
