% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.r
\name{plot.scatter}
\alias{plot.scatter}
\title{Plot Scatter of Compensated and Uncompensated Data}
\usage{
\method{plot}{scatter}(
  expr.data.unco.x,
  expr.data.unco.y,
  expr.data.comp.x,
  expr.data.comp.y,
  spillover.unco.inte,
  spillover.unco.coef,
  spillover.comp.inte,
  spillover.comp.coef,
  spillover.comp.slop,
  marker.limit.range,
  marker.proper.limit.range,
  samp,
  marker,
  marker.proper,
  scale.untransformed,
  figure.file.label,
  flow.gate,
  flow.control,
  asp
)
}
\arguments{
\item{expr.data.unco.x}{Numeric vector of uncompensated data for the x-axis.}

\item{expr.data.unco.y}{Numeric vector of uncompensated data for the y-axis.}

\item{expr.data.comp.x}{Numeric vector of compensated data for the x-axis.}

\item{expr.data.comp.y}{Numeric vector of compensated data for the y-axis.}

\item{spillover.unco.inte}{Numeric value of the intercept for the uncompensated
spillover.}

\item{spillover.unco.coef}{Numeric value of the coefficient for the uncompensated
spillover.}

\item{spillover.comp.inte}{Numeric value of the intercept for the compensated
spillover.}

\item{spillover.comp.coef}{Numeric value of the coefficient for the
compensated spillover.}

\item{spillover.comp.slop}{Numeric value of the slope error for the
compensated spillover.}

\item{marker.limit.range}{Numeric vector defining the limits for the marker range.}

\item{marker.proper.limit.range}{Numeric vector defining the limits for the
proper marker range.}

\item{samp}{Character string representing the sample identifier.}

\item{marker}{Character string representing the marker.}

\item{marker.proper}{Character string representing the proper marker.}

\item{scale.untransformed}{Logical value indicating whether to use
untransformed scale.}

\item{figure.file.label}{Character string representing the label for the
figure file.}

\item{flow.gate}{List containing flow gate parameters.}

\item{flow.control}{List containing flow cytometry control parameters.}

\item{asp}{List containing aesthetic parameters for the plot.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function creates scatter plots of compensated and
uncompensated data, including regression and reference lines, and a label
showing slope error.
}
